weedmap_inference/requirements.txt
gradio==4.44.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
transformers>=4.35.0
huggingface-hub>=0.19.0
scipy>=1.10.0